Completing info at DetailsView, adding region images, minor fixes and styles

+++ app/src/main/java/com/carlosjimz87/funwithflags/network/models/CountryDetails.kt
@@ -18,7 +18,7 @@ data class CountryDetails(
     val name: String,
     val nativeName: String?,
     val numericCode: String?,
-    val population: Int?,
+    val population: Long?,
     val region: String?,
     val subregion: String?,
     val timezones: List<String>,
